# backend/api/gen/go/admin/goods_category_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             (unknown)
// source: admin/goods_category.proto

package admin

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
	emptypb "google.golang.org/protobuf/types/known/emptypb"
	wrapperspb "google.golang.org/protobuf/types/known/wrapperspb"
	common "shop/api/gen/go/common"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationGoodsCategoryServiceCreateGoodsCategory = "/admin.GoodsCategoryService/CreateGoodsCategory"
const OperationGoodsCategoryServiceDeleteGoodsCategory = "/admin.GoodsCategoryService/DeleteGoodsCategory"
const OperationGoodsCategoryServiceGetGoodsCategory = "/admin.GoodsCategoryService/GetGoodsCategory"
const OperationGoodsCategoryServiceOptionGoodsCategory = "/admin.GoodsCategoryService/OptionGoodsCategory"
const OperationGoodsCategoryServiceSetGoodsCategoryStatus = "/admin.GoodsCategoryService/SetGoodsCategoryStatus"
const OperationGoodsCategoryServiceTreeGoodsCategory = "/admin.GoodsCategoryService/TreeGoodsCategory"
const OperationGoodsCategoryServiceUpdateGoodsCategory = "/admin.GoodsCategoryService/UpdateGoodsCategory"

type GoodsCategoryServiceHTTPServer interface {
	// CreateGoodsCategory 创建商品分类
	CreateGoodsCategory(context.Context, *GoodsCategoryForm) (*emptypb.Empty, error)
	// DeleteGoodsCategory 删除商品分类
	DeleteGoodsCategory(context.Context, *wrapperspb.StringValue) (*emptypb.Empty, error)
	// GetGoodsCategory 查询商品分类
	GetGoodsCategory(context.Context, *wrapperspb.Int64Value) (*GoodsCategoryForm, error)
	// OptionGoodsCategory 查询商品分类树形选择
	OptionGoodsCategory(context.Context, *OptionGoodsCategoryRequest) (*common.TreeOptionResponse, error)
	// SetGoodsCategoryStatus 设置状态
	SetGoodsCategoryStatus(context.Context, *common.SetStatusRequest) (*emptypb.Empty, error)
	// TreeGoodsCategory 查询商品分类树形列表
	TreeGoodsCategory(context.Context, *emptypb.Empty) (*TreeGoodsCategoryResponse, error)
	// UpdateGoodsCategory 更新商品分类
	UpdateGoodsCategory(context.Context, *GoodsCategoryForm) (*emptypb.Empty, error)
}

func RegisterGoodsCategoryServiceHTTPServer(s *http.Server, srv GoodsCategoryServiceHTTPServer) {
	r := s.Route("/")
	r.GET("/api/admin/goods/category/tree", _GoodsCategoryService_TreeGoodsCategory0_HTTP_Handler(srv))
	r.GET("/api/admin/goods/category/option", _GoodsCategoryService_OptionGoodsCategory0_HTTP_Handler(srv))
	r.GET("/api/admin/goods/category/{value}", _GoodsCategoryService_GetGoodsCategory0_HTTP_Handler(srv))
	r.POST("/api/admin/goods/category", _GoodsCategoryService_CreateGoodsCategory0_HTTP_Handler(srv))
	r.PUT("/api/admin/goods/category/{id}", _GoodsCategoryService_UpdateGoodsCategory0_HTTP_Handler(srv))
	r.DELETE("/api/admin/goods/category/{value}", _GoodsCategoryService_DeleteGoodsCategory0_HTTP_Handler(srv))
	r.PUT("/api/admin/goods/category/{id}/status", _GoodsCategoryService_SetGoodsCategoryStatus0_HTTP_Handler(srv))
}

func _GoodsCategoryService_TreeGoodsCategory0_HTTP_Handler(srv GoodsCategoryServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in emptypb.Empty
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationGoodsCategoryServiceTreeGoodsCategory)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.TreeGoodsCategory(ctx, req.(*emptypb.Empty))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*TreeGoodsCategoryResponse)
		return ctx.Result(200, reply)
	}
}

func _GoodsCategoryService_OptionGoodsCategory0_HTTP_Handler(srv GoodsCategoryServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in OptionGoodsCategoryRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationGoodsCategoryServiceOptionGoodsCategory)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.OptionGoodsCategory(ctx, req.(*OptionGoodsCategoryRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*common.TreeOptionResponse)
		return ctx.Result(200, reply)
	}
}

func _GoodsCategoryService_GetGoodsCategory0_HTTP_Handler(srv GoodsCategoryServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in wrapperspb.Int64Value
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationGoodsCategoryServiceGetGoodsCategory)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetGoodsCategory(ctx, req.(*wrapperspb.Int64Value))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GoodsCategoryForm)
		return ctx.Result(200, reply)
	}
}

func _GoodsCategoryService_CreateGoodsCategory0_HTTP_Handler(srv GoodsCategoryServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GoodsCategoryForm
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationGoodsCategoryServiceCreateGoodsCategory)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.CreateGoodsCategory(ctx, req.(*GoodsCategoryForm))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*emptypb.Empty)
		return ctx.Result(200, reply)
	}
}

func _GoodsCategoryService_UpdateGoodsCategory0_HTTP_Handler(srv GoodsCategoryServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GoodsCategoryForm
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationGoodsCategoryServiceUpdateGoodsCategory)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.UpdateGoodsCategory(ctx, req.(*GoodsCategoryForm))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*emptypb.Empty)
		return ctx.Result(200, reply)
	}
}

func _GoodsCategoryService_DeleteGoodsCategory0_HTTP_Handler(srv GoodsCategoryServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in wrapperspb.StringValue
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationGoodsCategoryServiceDeleteGoodsCategory)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.DeleteGoodsCategory(ctx, req.(*wrapperspb.StringValue))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*emptypb.Empty)
		return ctx.Result(200, reply)
	}
}

func _GoodsCategoryService_SetGoodsCategoryStatus0_HTTP_Handler(srv GoodsCategoryServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in common.SetStatusRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationGoodsCategoryServiceSetGoodsCategoryStatus)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.SetGoodsCategoryStatus(ctx, req.(*common.SetStatusRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*emptypb.Empty)
		return ctx.Result(200, reply)
	}
}

type GoodsCategoryServiceHTTPClient interface {
	// CreateGoodsCategory 创建商品分类
	CreateGoodsCategory(ctx context.Context, req *GoodsCategoryForm, opts ...http.CallOption) (rsp *emptypb.Empty, err error)
	// DeleteGoodsCategory 删除商品分类
	DeleteGoodsCategory(ctx context.Context, req *wrapperspb.StringValue, opts ...http.CallOption) (rsp *emptypb.Empty, err error)
	// GetGoodsCategory 查询商品分类
	GetGoodsCategory(ctx context.Context, req *wrapperspb.Int64Value, opts ...http.CallOption) (rsp *GoodsCategoryForm, err error)
	// OptionGoodsCategory 查询商品分类树形选择
	OptionGoodsCategory(ctx context.Context, req *OptionGoodsCategoryRequest, opts ...http.CallOption) (rsp *common.TreeOptionResponse, err error)
	// SetGoodsCategoryStatus 设置状态
	SetGoodsCategoryStatus(ctx context.Context, req *common.SetStatusRequest, opts ...http.CallOption) (rsp *emptypb.Empty, err error)
	// TreeGoodsCategory 查询商品分类树形列表
	TreeGoodsCategory(ctx context.Context, req *emptypb.Empty, opts ...http.CallOption) (rsp *TreeGoodsCategoryResponse, err error)
	// UpdateGoodsCategory 更新商品分类
	UpdateGoodsCategory(ctx context.Context, req *GoodsCategoryForm, opts ...http.CallOption) (rsp *emptypb.Empty, err error)
}

type GoodsCategoryServiceHTTPClientImpl struct {
	cc *http.Client
}

func NewGoodsCategoryServiceHTTPClient(client *http.Client) GoodsCategoryServiceHTTPClient {
	return &GoodsCategoryServiceHTTPClientImpl{client}
}

// CreateGoodsCategory 创建商品分类
func (c *GoodsCategoryServiceHTTPClientImpl) CreateGoodsCategory(ctx context.Context, in *GoodsCategoryForm, opts ...http.CallOption) (*emptypb.Empty, error) {
	var out emptypb.Empty
	pattern := "/api/admin/goods/category"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationGoodsCategoryServiceCreateGoodsCategory))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// DeleteGoodsCategory 删除商品分类
func (c *GoodsCategoryServiceHTTPClientImpl) DeleteGoodsCategory(ctx context.Context, in *wrapperspb.StringValue, opts ...http.CallOption) (*emptypb.Empty, error) {
	var out emptypb.Empty
	pattern := "/api/admin/goods/category/{value}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationGoodsCategoryServiceDeleteGoodsCategory))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "DELETE", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// GetGoodsCategory 查询商品分类
func (c *GoodsCategoryServiceHTTPClientImpl) GetGoodsCategory(ctx context.Context, in *wrapperspb.Int64Value, opts ...http.CallOption) (*GoodsCategoryForm, error) {
	var out GoodsCategoryForm
	pattern := "/api/admin/goods/category/{value}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationGoodsCategoryServiceGetGoodsCategory))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// OptionGoodsCategory 查询商品分类树形选择
func (c *GoodsCategoryServiceHTTPClientImpl) OptionGoodsCategory(ctx context.Context, in *OptionGoodsCategoryRequest, opts ...http.CallOption) (*common.TreeOptionResponse, error) {
	var out common.TreeOptionResponse
	pattern := "/api/admin/goods/category/option"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationGoodsCategoryServiceOptionGoodsCategory))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// SetGoodsCategoryStatus 设置状态
func (c *GoodsCategoryServiceHTTPClientImpl) SetGoodsCategoryStatus(ctx context.Context, in *common.SetStatusRequest, opts ...http.CallOption) (*emptypb.Empty, error) {
	var out emptypb.Empty
	pattern := "/api/admin/goods/category/{id}/status"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationGoodsCategoryServiceSetGoodsCategoryStatus))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PUT", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// TreeGoodsCategory 查询商品分类树形列表
func (c *GoodsCategoryServiceHTTPClientImpl) TreeGoodsCategory(ctx context.Context, in *emptypb.Empty, opts ...http.CallOption) (*TreeGoodsCategoryResponse, error) {
	var out TreeGoodsCategoryResponse
	pattern := "/api/admin/goods/category/tree"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationGoodsCategoryServiceTreeGoodsCategory))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// UpdateGoodsCategory 更新商品分类
func (c *GoodsCategoryServiceHTTPClientImpl) UpdateGoodsCategory(ctx context.Context, in *GoodsCategoryForm, opts ...http.CallOption) (*emptypb.Empty, error) {
	var out emptypb.Empty
	pattern := "/api/admin/goods/category/{id}"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationGoodsCategoryServiceUpdateGoodsCategory))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PUT", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}