# backend/internal/cmd/server/assets/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /api/admin/goods/category:
        post:
            tags:
                - GoodsCategoryService
            description: 创建商品分类
            operationId: GoodsCategoryService_CreateGoodsCategory
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/GoodsCategoryForm'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
    /api/admin/goods/category/option:
        get:
            tags:
                - GoodsCategoryService
            description: 查询商品分类树形选择
            operationId: GoodsCategoryService_OptionGoodsCategory
            parameters:
                - name: parentId
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/TreeOptionResponse'
...
    /api/admin/goods/category/{id}:
        put:
            tags:
                - GoodsCategoryService
            description: 更新商品分类
            operationId: GoodsCategoryService_UpdateGoodsCategory
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/GoodsCategoryForm'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
    /api/admin/goods/category/{id}/status:
        put:
            tags:
                - GoodsCategoryService
            description: 设置状态
            operationId: GoodsCategoryService_SetGoodsCategoryStatus
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/SetStatusRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
    /api/admin/goods/category/{value}:
        get:
            tags:
                - GoodsCategoryService
            description: 查询商品分类
            operationId: GoodsCategoryService_GetGoodsCategory
            parameters:
                - name: value
                  in: path
                  description: The int64 value.
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GoodsCategoryForm'
        delete:
            tags:
                - GoodsCategoryService
            description: 删除商品分类
            operationId: GoodsCategoryService_DeleteGoodsCategory
            parameters:
                - name: value
                  in: path
                  description: The string value.
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content: {}
...
    /api/app/goods/category:
        get:
            tags:
                - GoodsCategoryService
            description: 查询商品分类列表
            operationId: GoodsCategoryService_ListGoodsCategory
            parameters:
                - name: parentId
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListGoodsCategoryResponse'
...
    /api/app/shop/hot/goods:
        get:
            tags:
                - ShopHotService
            description: 查询热门推荐商品
            operationId: ShopHotService_PageShopHotGoods
            parameters:
                - name: hotItemId
                  in: query
                  description: 选项id
                  schema:
                    type: string
                - name: pageNum
                  in: query
                  description: 当前页码
                  schema:
                    type: string
                - name: pageSize
                  in: query
                  description: 每一页的行数
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/PageShopHotGoodsResponse'
...
components:
    schemas:
...
        GoodsCategory:
            type: object
            properties:
                id:
                    type: string
                    description: 商品分类ID
                parentId:
                    type: string
                    description: 父级商品分类ID
                name:
                    type: string
                    description: 商品分类名称
                picture:
                    type: string
                    description: 商品分类图片
                sort:
                    type: integer
                    description: 排序
                    format: int32
                status:
                    enum:
                        - UNKNOWN
                        - ENABLE
                        - DISABLE
                    type: string
                    description: 菜单状态
                    format: enum
                createdAt:
                    type: string
                    description: 创建时间
                updatedAt:
                    type: string
                    description: 更新时间
                children:
                    type: array
                    items:
                        $ref: '#/components/schemas/GoodsCategory'
                    description: 子节点树
            description: 商品分类
        GoodsCategoryForm:
            type: object
            properties:
                id:
                    type: string
                    description: 商品分类ID
                parentId:
                    type: string
                    description: 父级商品分类ID
                name:
                    type: string
                    description: 商品分类名称
                picture:
                    type: string
                    description: 商品分类图片
                sort:
                    type: integer
                    description: 排序
                    format: int32
                status:
                    enum:
                        - UNKNOWN
                        - ENABLE
                        - DISABLE
                    type: string
                    description: 菜单状态
                    format: enum
            description: 商品分类
        GoodsInfo:
            type: object
            properties:
                id:
                    type: string
                    description: 商品ID
                categoryId:
                    type: string
                    description: 分类ID
                name:
                    type: string
                    description: 名称
                desc:
                    type: string
                    description: 描述
                picture:
                    type: string
                    description: 商品图片
                price:
                    type: string
                    description: 当前价格(分)
                discountPrice:
                    type: string
                    description: 折扣价格（分）
                initSaleNum:
                    type: string
                    description: 初始销量
                realSaleNum:
                    type: string
                    description: 真实销售数量
                inventory:
                    type: string
                    description: 库存数量
                status:
                    enum:
                        - UNKNOWN_GS
                        - PUT_ON
                        - PULL_OFF
                    type: string
                    description: 状态
                    format: enum
                createdAt:
                    type: string
                    description: 创建时间
                updatedAt:
                    type: string
                    description: 更新时间
                categoryName:
                    type: string
                    description: 分类名称
...
        ListGoodsCategoryResponse:
            type: object
            properties:
                list:
                    type: array
                    items:
                        $ref: '#/components/schemas/GoodsCategory'
                    description: 分页数据
...
        PageShopHotGoodsResponse:
            type: object
            properties:
                list:
                    type: array
                    items:
                        $ref: '#/components/schemas/GoodsInfo'
                    description: 分页数据
                total:
                    type: integer
                    description: 总数
                    format: int32
...
        SetStatusRequest:
            type: object
            properties:
                id:
                    type: string
                    description: 主键id
                status:
                    type: integer
                    description: 状态
                    format: int32
...
        TreeOptionResponse:
            type: object
            properties:
                list:
                    type: array
                    items:
                        $ref: '#/components/schemas/TreeOptionResponse_Option'
        TreeOptionResponse_Option:
            type: object
            properties:
                label:
                    type: string
                    description: 选项名
                value:
                    type: string
                    description: 选项值
                disabled:
                    type: boolean
                    description: 是否禁用
                children:
                    type: array
                    items:
                        $ref: '#/components/schemas/TreeOptionResponse_Option'
                    description: 子节点树
...